Modularizes linting utilities for reusability
Refactors linting logic by extracting helper functions into dedicated modules for colors and components.

Improves code organization and maintainability by centralizing reusable utilities like `findFillStyleNodes`, `canNodeHaveFillStyle`, and `isUsingPaintStyle`.

Introduces new helper modules for Astro components to streamline component loading and validation.
