+++ src/lint/colors/index.ts
@@ -1,0 +1,4 @@
+import {findFillStyleNodes, canNodeHaveFillStyle} from "./helpers";
+import { isUsingPaintStyle } from "./tests";
+
+export {findFillStyleNodes, canNodeHaveFillStyle, isUsingPaintStyle};
