+++ src/lint/components/types.ts
@@ -1,0 +1,6 @@
+interface AstroComponent {
+  type: "COMPONENT" | "COMPONENT_SET";
+  key: string;
+}
+
+export type { AstroComponent };
